Add to detection usecase dump history and custom signal labelling
